+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.1.97 (2023-10-16)
 
+- fix: set default model for project
+
 ## 0.1.96 (2023-10-16)
 
 - feat: project settings page